requirements.txt
fastapi 

uvicorn[standard]

sqlalchemy

asyncpg

minio

fastapi-users[sqlalchemy]

httpx 

pillow

python-dotenv

python-multipart

alembic

pyJWT

pwdlib

celery

langchain

langchain-community

chromadb

google-genai

redis